Test Boss Implementation
As a test - I have made a boss sprite that comes in from the side using the animator. Then is able to generate text in a speech buble the game. It takes dialogue as in an imput. This enables us to call it from main, or where ever to drive the narrative. I will show this off to the team tomorrow.